Add board subcomponent
