+++ src/app/content/memoboard/memoboard.component.css
@@ -7,6 +7,7 @@
     background-color: white;
     width: 100%;
     padding: 10px;
+    min-height: 40px;
 }
 
 .kanban-header .top-bar mat-icon {
@@ -39,100 +40,10 @@
     font-weight: 450;
 }
 
-.boards-wrapper {
-    height: 100%;
-}
-
-.board {
-    width: 350px;
-    border-radius: 10px;
-}
-
-.board-header {
-    background: white;
-    border-radius: 10px;
-    height: 50px;
-    padding: 0 15px;
-}
-
-.board-header eunes-button ::ng-deep .background {
-    background: aliceblue;
-}
-
-.board-top-bar {
-    padding: 0px 10px;
-}
-
-.memos {
-    border-radius: 10px;
-    padding: 10px;
-    /* background: rgba(211, 211, 211, 0.5); */
-}
-
-.memo {
-    background: white;
-    border-radius: 10px;
-    padding: 15px;
-    margin: 8px 0;
-    cursor: pointer;
-    box-shadow: 0px 1px 3px lightsteelblue;
-}
-
-.memo-header .title {
-    font-weight: bold;
-}
-
-textarea {
-    outline: none;
-    background-color: transparent;
-    padding: 5px;
-    resize: none;
+::ng-deep .project-dialog > mat-dialog-container {
+    box-shadow: 0px 1px 10px lightsteelblue, 0 0 0 1px lightsteelblue;
 }
 
 eunes-button ::ng-deep .background {
     background-color: white;
-}
-
-.cdk-drag-preview {
-    box-sizing: border-box;
-    box-shadow: 0 5px 5px -3px rgba(0, 0, 0, 0.2),
-                0 8px 10px 1px rgba(0, 0, 0, 0.14),
-                0 3px 14px 2px rgba(0, 0, 0, 0.12);
-}
-  
-.cdk-drag-placeholder {
-    background: lightsteelblue;
-    opacity: 0.2;
-    border: solid 1px lightslategrey;
-    transition: transform 250ms cubic-bezier(0, 0, 0.2, 1);
-}
-
-.cdk-drag-placeholder > * {
-    visibility: hidden;
-}
-
-.cdk-drag-animating {
-    transition: transform 250ms cubic-bezier(0, 0, 0.2, 1);
-}
-
-.memo:last-child {
-    border: none;
-}
-
-.memos.cdk-drop-list-dragging .memo:not(.cdk-drag-placeholder) {
-    transition: transform 250ms cubic-bezier(0, 0, 0.2, 1);
-}
-
-.memos.cdk-drop-list-dragging {
-    background: rgba(245, 245, 245, 0.5);
-    border: dashed 1px lightgray;
-    background-size: 120%;
-}
-
-.new-board eunes-button ::ng-deep .background {
-    background: aliceblue;
-}
-
-::ng-deep .project-dialog > mat-dialog-container {
-    box-shadow: 0px 1px 10px lightsteelblue, 0 0 0 1px lightsteelblue;
 }
