+++ src/app/elements/button/button.component.css
@@ -27,7 +27,9 @@ button:hover {
 button .text {
     transition-duration: 3s;
     margin-top: 1px;
-    margin-left: 5px;
+    /* margin-left: 5px; */
+    font-size: 14px;
+    font-weight: 500;
 }
 
 button:active {
